@@ -5,21 +5,20 @@ import '../providers/Products.dart';
 
 class ProductDetail extends StatelessWidget {
   static const routeName = '/product';
-  int productId;
 
   @override
   Widget build(BuildContext context) {
-    this.productId = ModalRoute.of(context).settings.arguments;
-    final product = Provider.of<Products>(context).findById(this.productId);
+    final productId = ModalRoute.of(context)!.settings.arguments as int;
+    final product = Provider.of<Products>(context).findById(productId);
 
     return Scaffold(
       appBar: AppBar(
-        title: Text(product.title),
+        title: Text(product.title!),
       ),
       body: Column(
         children: [
-          Image.network(product.imageUrl, fit: BoxFit.contain),
-          Text(product.description),
+          Image.network(product.imageUrl!, fit: BoxFit.contain),
+          Text(product.description!),
         ],
       ),
     );
